# Conflicts:
#	services/distribution/src/main/java/app/coronawarn/server/services/distribution/assembly/appconfig/ApplicationConfigurationV2PublicationConfig.java
#	services/distribution/src/main/java/app/coronawarn/server/services/distribution/assembly/tracewarnings/TraceTimeIntervalWarningsPackageBundler.java
#	services/distribution/src/main/java/app/coronawarn/server/services/distribution/config/DistributionServiceConfig.java
#	services/distribution/src/main/resources/application.yaml
#	services/distribution/src/main/resources/main-config/v2/plausible-deniability-parameters.yaml
#	services/distribution/src/main/resources/main-config/v2/presence-tracing-parameters.yaml
#	services/distribution/src/test/java/app/coronawarn/server/services/distribution/assembly/appconfig/parsing/v2/DeserializedPlausibleDeniabilityParametersTest.java
#	services/distribution/src/test/resources/application.yaml

+++ services/distribution/src/main/java/app/coronawarn/server/services/distribution/assembly/component/QrCodePosterTemplateStructureProvider.java
@@ -65,12 +65,12 @@ public class QrCodePosterTemplateStructureProvider {
   private QRCodePosterTemplateAndroid buildAndroidProtoStructure(
       QrCodePosterTemplate templateConfig) {
     DescriptionTextBox textBoxConfig = templateConfig.getDescriptionTextBox();
-    return QRCodePosterTemplateAndroid.newBuilder().setOffsetX(templateConfig.getOffsetX())
-        .setOffsetY(templateConfig.getOffsetY())
+    return QRCodePosterTemplateAndroid.newBuilder().setOffsetX(templateConfig.getOffsetX().floatValue())
+        .setOffsetY(templateConfig.getOffsetY().floatValue())
         .setTemplate(qrTemplateLoader.loadAndroidTemplateAsBytes())
         .setQrCodeSideLength(templateConfig.getQrCodeSideLength())
         .setDescriptionTextBox(QRCodeTextBoxAndroid.newBuilder()
-            .setOffsetX(textBoxConfig.getOffsetX()).setOffsetY(textBoxConfig.getOffsetY())
+            .setOffsetX(textBoxConfig.getOffsetX().floatValue()).setOffsetY(textBoxConfig.getOffsetY().floatValue())
             .setFontSize(textBoxConfig.getFontSize()).setHeight(textBoxConfig.getHeight())
             .setFontColor(textBoxConfig.getFontColor()).build())
         .build();
@@ -78,12 +78,12 @@ public class QrCodePosterTemplateStructureProvider {
 
   private QRCodePosterTemplateIOS buildIosProtoStructure(QrCodePosterTemplate templateConfig) {
     DescriptionTextBox textBoxConfig = templateConfig.getDescriptionTextBox();
-    return QRCodePosterTemplateIOS.newBuilder().setOffsetX(templateConfig.getOffsetX())
-        .setOffsetY(templateConfig.getOffsetY())
+    return QRCodePosterTemplateIOS.newBuilder().setOffsetX(templateConfig.getOffsetX().intValue())
+        .setOffsetY(templateConfig.getOffsetY().intValue())
         .setTemplate(qrTemplateLoader.loadIosTemplateAsBytes())
         .setQrCodeSideLength(templateConfig.getQrCodeSideLength())
-        .setDescriptionTextBox(QRCodeTextBoxIOS.newBuilder().setOffsetX(textBoxConfig.getOffsetX())
-            .setOffsetY(textBoxConfig.getOffsetY()).setFontSize(textBoxConfig.getFontSize())
+        .setDescriptionTextBox(QRCodeTextBoxIOS.newBuilder().setOffsetX(textBoxConfig.getOffsetX().intValue())
+            .setOffsetY(textBoxConfig.getOffsetY().intValue()).setFontSize(textBoxConfig.getFontSize())
             .setHeight(textBoxConfig.getHeight()).setFontColor(textBoxConfig.getFontColor())
             .build())
         .build();
